+++ CHANGELOG.md
@@ -157,3 +157,7 @@
 
 ## [v4.7.1](https://github.com/xhdlphzr/FranxAI/releases/tag/v4.7.1)
 - Added `.gitignore` file.
+
+## [v4.8.0](https://github.com/xhdlphzr/FranxAI/releases/tag/v4.8.0)
+- Added full-platform shared conversation history feature.
+- Fixed partial response save issue.